RVNGTextInterface: document new libepubgen fixed layout key
LibreOffice produces this and libepubgen understands it. It's for the case when
the document is a series of SVG images (to preserve the layout of the document
as-is), but still the generator want to produce a nice table of contents, so it
wants to list chapters and have them point to the relevant page. This key
allows the setup where a page span always contains a single page, so simply a
string vector is enough as a property.